all get requests operating correctly for both db's
