+++ server/OldServerRoutes/drawers.js
@@ -1,0 +1,9 @@
+const Router = require("express-promise-router");
+
+const router = new Router();
+
+module.exports = router;
+
+router.get("/", async (req, res) => {
+  res.send();
+});
