+++ server/db/data/CSVwriter.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 
-const combinedGenerator = require('./combinedGenerator');
+// const combinedGenerator = require('./combinedGenerator');
 
 let randomStart = 3;
 let targetCount = 10000000;
