--- server/routes/index.js
@@ -1,9 +1,0 @@
-const overview = require("./overview");
-const specs = require("./specs");
-const reviews = require("./reviews");
-
-module.exports = app => {
-    app.use('/overview', overview);
-    app.use('/specs', specs);
-    app.use('/reviews', reviews);
-}
